# lib/api-zod/src/generated/types/createIssueInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CivicReport API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateIssueInput {
  title: string;
  description?: string;
  category: string;
  address: string;
  constituency?: string;
  city?: string;
  state?: string;
  lat: number;
  lng: number;
  photoUrl?: string;
  urgent?: boolean;
  anonymous?: boolean;
}