# lib/api-zod/src/generated/types/notification.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CivicReport API
 * OpenAPI spec version: 0.1.0
 */

export interface Notification {
  id: string;
  message: string;
  type?: string;
  issueId?: string;
  createdAt: string;
  read: boolean;
}