# lib/api-zod/src/generated/types/publicStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CivicReport API
 * OpenAPI spec version: 0.1.0
 */

export interface PublicStats {
  totalReported: number;
  totalVerified: number;
  totalResolved: number;
}